feat: move RetrievalDialog to common components instead of Timeline components
