+++ components/timeline/index.ts
@@ -1,2 +1,1 @@
 export * from "./CoreTimeline";
-export * from "./RetrievalDialog";
